# src/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TableBook API",
    "author": "Alban Okoby",
    "description": "REST API for a restaurant table reservation platform. Users browse restaurants and make reservations; restaurant owners manage their listings and bookings; admins manage users and approve restaurants.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/", "description": "Current host" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "details": { "type": "array", "items": { "type": "object" } }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "username": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "phone": { "type": "string", "nullable": true },
          "avatar": { "type": "string", "nullable": true },
          "bio": { "type": "string" },
          "role": { "type": "string", "enum": ["user", "restaurant_owner", "admin"] },
          "isVerified": { "type": "boolean" },
          "isActive": { "type": "boolean" },
          "lastLogin": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Table": {
        "type": "object",
        "properties": {
          "label": { "type": "string" },
          "capacity": { "type": "integer", "minimum": 1, "maximum": 20 },
          "count": { "type": "integer", "minimum": 1 }
        }
      },
      "OpeningHour": {
        "type": "object",
        "properties": {
          "day": { "type": "string", "enum": ["monday", "tuesday", "wednesday", "thursday", "friday", "saturday", "sunday"] },
          "open": { "type": "string", "example": "12:00" },
          "close": { "type": "string", "example": "23:00" },
          "isClosed": { "type": "boolean" }
        }
      },
      "Review": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "user": { "type": "string", "description": "User ID (populated with username/avatar on read)" },
          "rating": { "type": "integer", "minimum": 1, "maximum": 5 },
          "title": { "type": "string" },
          "body": { "type": "string" },
          "likes": { "type": "array", "items": { "type": "string" } },
          "isEdited": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "Address": {
        "type": "object",
        "required": ["city"],
        "properties": {
          "street": { "type": "string" },
          "city": { "type": "string" },
          "state": { "type": "string" },
          "country": { "type": "string", "default": "US" },
          "zipCode": { "type": "string" }
        }
      },
      "Restaurant": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "cuisine": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["Italian", "French", "Japanese", "Chinese", "Indian", "Mexican", "American", "Mediterranean", "Thai", "Greek", "Spanish", "Lebanese", "Korean", "Vietnamese", "Turkish", "Moroccan", "Brazilian", "Seafood", "Vegetarian", "Vegan", "Steakhouse", "Fusion", "Other"]
            }
          },
          "address": { "$ref": "#/components/schemas/Address" },
          "phone": { "type": "string" },
          "email": { "type": "string" },
          "website": { "type": "string" },
          "images": { "type": "array", "items": { "type": "string" } },
          "coverImage": { "type": "string", "nullable": true },
          "priceRange": { "type": "string", "enum": ["$", "$$", "$$$", "$$$$"] },
          "tables": { "type": "array", "items": { "$ref": "#/components/schemas/Table" } },
          "openingHours": { "type": "array", "items": { "$ref": "#/components/schemas/OpeningHour" } },
          "ratings": {
            "type": "object",
            "properties": {
              "average": { "type": "number" },
              "count": { "type": "integer" }
            }
          },
          "reviews": { "type": "array", "items": { "$ref": "#/components/schemas/Review" } },
          "addedBy": { "type": "string", "description": "User ID of the owner/admin who added it" },
          "isFeatured": { "type": "boolean" },
          "isApproved": { "type": "boolean" },
          "viewCount": { "type": "integer" },
          "reviewCount": { "type": "integer", "readOnly": true },
          "totalCapacity": { "type": "integer", "readOnly": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Booking": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "restaurant": { "type": "string", "description": "Restaurant ID (populated on read)" },
          "user": { "type": "string", "description": "User ID (populated on read)" },
          "date": { "type": "string", "format": "date-time" },
          "time": { "type": "string", "example": "19:30" },
          "partySize": { "type": "integer", "minimum": 1, "maximum": 50 },
          "tableLabel": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["pending", "confirmed", "cancelled", "completed", "no-show"] },
          "specialRequests": { "type": "string" },
          "confirmationCode": { "type": "string" },
          "cancelledAt": { "type": "string", "format": "date-time", "nullable": true },
          "cancelledBy": { "type": "string", "nullable": true },
          "cancelReason": { "type": "string" },
          "isPast": { "type": "boolean", "readOnly": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "pages": { "type": "integer" }
        }
      }
    },
    "responses": {
      "ValidationError": {
        "description": "Validation failed",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Unauthorized": {
        "description": "Missing or invalid authentication token",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Forbidden": {
        "description": "Authenticated but not allowed to perform this action",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    }
  },
  "paths": {
...
    "/api/restaurants": {
      "get": {
        "tags": ["Restaurants"],
        "summary": "Search/list restaurants",
        "security": [],
        "parameters": [
          { "name": "q", "in": "query", "schema": { "type": "string" }, "description": "Full-text search query" },
          { "name": "cuisine", "in": "query", "schema": { "type": "string" }, "description": "Comma-separated cuisine list" },
          { "name": "city", "in": "query", "schema": { "type": "string" } },
          { "name": "priceRange", "in": "query", "schema": { "type": "string", "enum": ["$", "$$", "$$$", "$$$$"] } },
          { "name": "minRating", "in": "query", "schema": { "type": "number", "minimum": 0, "maximum": 5 } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string", "enum": ["relevance", "newest", "rating", "popular", "name"], "default": "newest" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 50, "default": 12 } }
        ],
        "responses": {
          "200": {
            "description": "Restaurant list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "restaurants": { "type": "array", "items": { "$ref": "#/components/schemas/Restaurant" } },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Restaurants"],
        "summary": "Create a restaurant",
        "description": "Requires role admin or restaurant_owner",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "cuisine", "address"],
                "properties": {
                  "name": { "type": "string", "maxLength": 200 },
                  "description": { "type": "string", "maxLength": 3000 },
                  "cuisine": { "type": "array", "minItems": 1, "items": { "type": "string" } },
                  "address": { "$ref": "#/components/schemas/Address" },
                  "phone": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "priceRange": { "type": "string", "enum": ["$", "$$", "$$$", "$$$$"] },
                  "tables": { "type": "array", "items": { "$ref": "#/components/schemas/Table" } },
                  "openingHours": { "type": "array", "items": { "$ref": "#/components/schemas/OpeningHour" } }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Restaurant created", "content": { "application/json": { "schema": { "type": "object", "properties": { "restaurant": { "$ref": "#/components/schemas/Restaurant" } } } } } },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
    "/api/restaurants/{id}": {
      "get": {
        "tags": ["Restaurants"],
        "summary": "Get a restaurant by ID",
        "security": [],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Restaurant", "content": { "application/json": { "schema": { "type": "object", "properties": { "restaurant": { "$ref": "#/components/schemas/Restaurant" } } } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["Restaurants"],
        "summary": "Update a restaurant",
        "description": "Owner of the restaurant or an admin",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Any subset of restaurant fields. Admins may also set isApproved/isFeatured; owners may set isApproved to show/hide their listing.",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "cuisine": { "type": "array", "items": { "type": "string" } },
                  "address": { "$ref": "#/components/schemas/Address" },
                  "phone": { "type": "string" },
                  "email": { "type": "string" },
                  "website": { "type": "string" },
                  "images": { "type": "array", "items": { "type": "string" } },
                  "coverImage": { "type": "string" },
                  "priceRange": { "type": "string", "enum": ["$", "$$", "$$$", "$$$$"] },
                  "tables": { "type": "array", "items": { "$ref": "#/components/schemas/Table" } },
                  "openingHours": { "type": "array", "items": { "$ref": "#/components/schemas/OpeningHour" } },
                  "isApproved": { "type": "boolean" },
                  "isFeatured": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated restaurant", "content": { "application/json": { "schema": { "type": "object", "properties": { "restaurant": { "$ref": "#/components/schemas/Restaurant" } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Restaurants"],
        "summary": "Delete a restaurant",
        "description": "Admin only",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Restaurant deleted", "content": { "application/json": { "schema": { "type": "object", "properties": { "message": { "type": "string" } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/api/bookings": {
      "get": {
        "tags": ["Bookings"],
        "summary": "List the current user's bookings",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["pending", "confirmed", "cancelled", "completed", "no-show"] } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 50, "default": 10 } }
        ],
        "responses": {
          "200": {
            "description": "Bookings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "bookings": { "type": "array", "items": { "$ref": "#/components/schemas/Booking" } },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Bookings"],
        "summary": "Create a reservation",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["restaurant", "date", "time", "partySize"],
                "properties": {
                  "restaurant": { "type": "string", "description": "Restaurant ID" },
                  "date": { "type": "string", "format": "date", "example": "2026-08-15" },
                  "time": { "type": "string", "example": "19:30" },
                  "partySize": { "type": "integer", "minimum": 1, "maximum": 50 },
                  "specialRequests": { "type": "string", "maxLength": 500 },
                  "tableLabel": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Booking created", "content": { "application/json": { "schema": { "type": "object", "properties": { "booking": { "$ref": "#/components/schemas/Booking" } } } } } },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/bookings/{id}": {
      "get": {
        "tags": ["Bookings"],
        "summary": "Get a booking by ID",
        "description": "Owner of the booking, an admin, or a restaurant_owner",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Booking", "content": { "application/json": { "schema": { "type": "object", "properties": { "booking": { "$ref": "#/components/schemas/Booking" } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List users",
        "description": "Admin only",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "q", "in": "query", "schema": { "type": "string" }, "description": "Search by username or email" },
          { "name": "role", "in": "query", "schema": { "type": "string", "enum": ["user", "restaurant_owner", "admin"] } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "User list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": { "type": "array", "items": { "$ref": "#/components/schemas/User" } },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get a user by ID",
        "security": [],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "User", "content": { "application/json": { "schema": { "type": "object", "properties": { "user": { "$ref": "#/components/schemas/User" } } } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  }
}